--- adapters/persistence/adapter-maria/src/main/java/com/sweet/domain/Photo.java
@@ -1,16 +1,0 @@
-package com.sweet.domain;
-
-import lombok.AccessLevel;
-import lombok.NoArgsConstructor;
-
-import javax.persistence.DiscriminatorValue;
-import javax.persistence.Entity;
-import javax.persistence.Table;
-
-@NoArgsConstructor(access = AccessLevel.PROTECTED)
-@Entity
-@Table(name = "photo")
-@DiscriminatorValue("photo")
-public class Photo extends Contents {
-
-}
