--- adapters/web/adapter-client-api/src/test/java/com/sweet/common/BaseControllerTest.java
@@ -1,28 +1,0 @@
-package com.sweet.common;
-
-import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
-
-import com.sweet.user.controller.SignUpController;
-import org.junit.jupiter.api.BeforeEach;
-import org.junit.jupiter.api.extension.ExtendWith;
-import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
-import org.springframework.test.context.junit.jupiter.SpringExtension;
-import org.springframework.test.web.servlet.MockMvc;
-import org.springframework.test.web.servlet.setup.MockMvcBuilders;
-import org.springframework.web.context.WebApplicationContext;
-import org.springframework.web.filter.CharacterEncodingFilter;
-
-@ExtendWith(SpringExtension.class)
-@WebMvcTest(controllers = {SignUpController.class})
-public abstract class BaseControllerTest {
-  protected MockMvc mockMvc;
-
-  @BeforeEach
-  void setUp(WebApplicationContext webApplicationContext) {
-    this.mockMvc = MockMvcBuilders
-        .webAppContextSetup(webApplicationContext)
-        .addFilters(new CharacterEncodingFilter("UTF-8", true))
-        .alwaysDo(print())
-        .build();
-  }
-}
