+++ applications/admin-api/src/main/java/com/caregiver/AdminApplication.java
@@ -1,0 +1,17 @@
+package com.caregiver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class AdminApplication {
+
+
+
+
+
+  public static void main(String[] args) {
+    SpringApplication.run(AdminApplication.class, args);
+  }
+
+}
