--- applications/admin-api/src/main/java/com/sweet/AdminApplication.java
@@ -1,17 +1,0 @@
-package com.sweet;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class AdminApplication {
-
-
-
-
-
-  public static void main(String[] args) {
-    SpringApplication.run(AdminApplication.class, args);
-  }
-
-}
