+++ applications/client-api/src/main/java/com/caregiver/ClientApplication.java
@@ -1,0 +1,13 @@
+package com.caregiver;
+
+import org.springframework.boot.SpringApplication;
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class ClientApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.run(ClientApplication.class, args);
+  }
+
+}
