--- applications/client-api/src/main/java/com/sweet/ClientApplication.java
@@ -1,13 +1,0 @@
-package com.sweet;
-
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-
-@SpringBootApplication
-public class ClientApplication {
-
-  public static void main(String[] args) {
-    SpringApplication.run(ClientApplication.class, args);
-  }
-
-}
