+++ common/src/main/java/com/caregiver/common/annotation/UseCase.java
@@ -1,0 +1,19 @@
+package com.caregiver.common.annotation;
+
+import java.lang.annotation.Documented;
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+import org.springframework.core.annotation.AliasFor;
+import org.springframework.stereotype.Component;
+
+@Target({ElementType.TYPE})
+@Retention(RetentionPolicy.RUNTIME)
+@Documented
+@Component
+public @interface UseCase {
+
+  @AliasFor(annotation = Component.class)
+  String value() default "";
+}
