+++ domain/src/main/java/com/caregiver/user/port/in/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * 유저 도메인 모듈 인입 UseCase package.
+ */
+
+package com.caregiver.user.port.in;
