+++ domain/src/main/java/com/caregiver/user/service/package-info.java
@@ -1,0 +1,5 @@
+/**
+ * 유저 도메인 UseCase 구현 package.
+ */
+
+package com.caregiver.user.service;
