--- domain/src/main/java/com/sweet/user/port/in/package-info.java
@@ -1,5 +1,0 @@
-/**
- * 유저 도메인 모듈 인입 UseCase package.
- */
-
-package com.sweet.user.port.in;
