--- domain/src/main/java/com/sweet/user/service/package-info.java
@@ -1,5 +1,0 @@
-/**
- * 유저 도메인 UseCase 구현 package.
- */
-
-package com.sweet.user.service;
